refactor: rename check package to checkers
