+++ src/main/kotlin/drift/checkers/SymbolCollector.kt
@@ -1,4 +1,4 @@
-package drift.check
+package drift.checkers
 
 import drift.ast.*
 import drift.ast.Function
